services/refundService/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests>=2.7.0
python-dotenv==1.0.0
pika
prometheus_client==0.14.1
python-jose[cryptography]>=3.3.0
typing